requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysqlclient==2.1.1
packaging==23.1
Pillow==9.4.0
pytz==2022.7.1
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==2.0.2
